crazy-rmi     계속 작업중... 언제 끝나런지 ㅋ
crazy-singleton
    move video 옵션 끄기 기능 추가

+++ crazy-storage/src/main/java/jk/kamoru/crazy/shop/schedule/Scheduling.java
@@ -89,6 +89,8 @@ public class Scheduling {
 				if (video.isExistVideoFileList()) {
 					File destination = video.getDelegatePathFile();
 					for (File file : video.getFileAll()) {
+						if (file == null || !file.isFile())
+							continue;
 						if (!destination.getPath().equals(file.getParentFile().getPath())) {
 							try {
 								FileUtils.moveFileToDirectory(file, destination, false);
